Rename propTypes file to camelCase
