+++ src/components/DisplayNames/index.native.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Text} from 'react-native';
-import {propTypes, defaultProps} from './DisplayNamesPropTypes';
+import {propTypes, defaultProps} from './displayNamesPropTypes';
 
 // As we don't have to show tooltips of the Native platform so we simply render the full display names list.
 const DisplayNames = ({
